+++ cpp/example_code/dynamodb/dynamodb_getting_started_scenario.cpp
@@ -48,48 +48,10 @@
 #include <aws/core/http/HttpClient.h>
 #include <fstream>
 
-#ifdef _HAS_ZLIB_
-#include <zlib.h>
-#endif // _HAS_ZLIB_
-
 #include <array>
 
 namespace AwsDoc {
-    namespace DynamoDB {
-        /**
-         * Constants used for DynamoDB table creation.
-         */
-        static const Aws::String MOVIE_TABLE_NAME("doc-example-table-movies");
-        static const Aws::String YEAR_KEY("year");
-        static const Aws::String TITLE_KEY("title");
-        static const Aws::String INFO_KEY("info");
-        static const Aws::String RATING_KEY("rating");
-        static const Aws::String PLOT_KEY("plot");
-        static const int PROVISIONED_THROUGHPUT_UNITS = 10;
-        static const Aws::String ALLOCATION_TAG("dynamodb_scenario");
-        static const int ASTERIX_FILL_WIDTH = 88;
-
-        //! Delete a DynamoDB table.
-        /*!
-          \sa deleteDynamoTable()
-          \param tableName: The DynamoDB table's name.
-          \param dynamoClient: A DynamoDB client.
-          \return bool: Function succeeded.
-        */
-        static bool deleteDynamoTable(const Aws::String &tableName,
-                                      const Aws::DynamoDB::DynamoDBClient &dynamoClient);
-
-        //! Query a newly created DynamoDB table until it is active.
-        /*!
-          \sa waitTableActive()
-          \param waitTableActive: The DynamoDB table's name.
-          \param dynamoClient: A DynamoDB client.
-          \return bool: Function succeeded.
-        */
-        static bool waitTableActive(const Aws::String &tableName,
-                                    const Aws::DynamoDB::DynamoDBClient &dynamoClient);
-
-        //! Convert an Aws JsonView object to a map of DynamoDB attribute values.
+    namespace DynamoDB {       //! Convert an Aws JsonView object to a map of DynamoDB attribute values.
         /*!
           \sa movieJsonViewToAttributeMap()
           \param jsonView: An Aws JsonView.
@@ -97,59 +59,6 @@ namespace AwsDoc {
         */
         Aws::Map<Aws::String, Aws::DynamoDB::Model::AttributeValue>
         movieJsonViewToAttributeMap(const Aws::Utils::Json::JsonView &jsonView);
-
-        //! Command line prompt/response utility function.
-        /*!
-         \\sa askQuestion()
-         \param string: A question prompt.
-         \param test: Test function for response.
-         \return Aws::String: User's response.
-         */
-        static Aws::String askQuestion(const Aws::String &string,
-                                       const std::function<bool(
-                                               Aws::String)> &test = [](
-                                               const Aws::String &) -> bool { return true; });
-
-        //! Command line prompt/response utility function for an integer result.
-        /*!
-         \sa askQuestionForInt()
-         \param string: A question prompt.
-         \return int: User's response.
-         */
-        int askQuestionForInt(const std::string &string);
-
-        //! Command line prompt/response utility function for a float result confined to
-        //! a range.
-        /*!
-         \sa askQuestionForFloatRange()
-         \param string: A question prompt.
-         \param low: Low inclusive.
-         \param high: High inclusive.
-         \return float: User's response.
-         */
-        float
-        askQuestionForFloatRange(const Aws::String &string, float low, float high);
-
-        //! Command line prompt/response utility function for an int result confined to
-        //! a range.
-        /*!
-         \sa askQuestionForIntRange()
-         \param string: A question prompt.
-         \param low: Low inclusive.
-         \param high: High inclusive.
-         \return int: User's response.
-         */
-        int askQuestionForIntRange(const Aws::String &string, int low,
-                                   int high);
-
-        //! Utility function to log movie attributes to std::cout.
-        /*!
-         \sa printMovieInfo()
-         \param movieMap: Map of DynamoDB attribute values.
-         \return void
-         */
-        void printMovieInfo(
-                const Aws::Map<Aws::String, Aws::DynamoDB::Model::AttributeValue> &movieMap);
 
         //! Download a JSON movie database file from the web and unzip the file.
         /*!
@@ -177,69 +86,6 @@ bool AwsDoc::DynamoDB::dynamodbGettingStartedScenario(
 
     Aws::DynamoDB::DynamoDBClient dynamoClient(clientConfiguration);
 
-    bool movieTableAlreadyExisted = false;
-
-    // 1. Create a table.
-    {
-        Aws::DynamoDB::Model::CreateTableRequest request;
-
-        Aws::DynamoDB::Model::AttributeDefinition yearAttributeDefinition;
-        yearAttributeDefinition.SetAttributeName(YEAR_KEY);
-        yearAttributeDefinition.SetAttributeType(
-                Aws::DynamoDB::Model::ScalarAttributeType::N);
-        request.AddAttributeDefinitions(yearAttributeDefinition);
-
-        Aws::DynamoDB::Model::AttributeDefinition titleAttributeDefinition;
-        yearAttributeDefinition.SetAttributeName(TITLE_KEY);
-        yearAttributeDefinition.SetAttributeType(
-                Aws::DynamoDB::Model::ScalarAttributeType::S);
-        request.AddAttributeDefinitions(yearAttributeDefinition);
-
-        Aws::DynamoDB::Model::KeySchemaElement yearKeySchema;
-        yearKeySchema.WithAttributeName(YEAR_KEY).WithKeyType(
-                Aws::DynamoDB::Model::KeyType::HASH);
-        request.AddKeySchema(yearKeySchema);
-
-        Aws::DynamoDB::Model::KeySchemaElement titleKeySchema;
-        yearKeySchema.WithAttributeName(TITLE_KEY).WithKeyType(
-                Aws::DynamoDB::Model::KeyType::RANGE);
-        request.AddKeySchema(yearKeySchema);
-
-        Aws::DynamoDB::Model::ProvisionedThroughput throughput;
-        throughput.WithReadCapacityUnits(
-                PROVISIONED_THROUGHPUT_UNITS).WithWriteCapacityUnits(
-                PROVISIONED_THROUGHPUT_UNITS);
-        request.SetProvisionedThroughput(throughput);
-        request.SetTableName(MOVIE_TABLE_NAME);
-
-        std::cout << "Creating table '" << MOVIE_TABLE_NAME << "'..." << std::endl;
-        const Aws::DynamoDB::Model::CreateTableOutcome &result = dynamoClient.CreateTable(
-                request);
-        if (!result.IsSuccess()) {
-            if (result.GetError().GetErrorType() ==
-                Aws::DynamoDB::DynamoDBErrors::RESOURCE_IN_USE) {
-                std::cout << "Table already exists." << std::endl;
-                movieTableAlreadyExisted = true;
-            }
-            else {
-                std::cerr << "Failed to create table: "
-                          << result.GetError().GetMessage();
-                return false;
-            }
-        }
-    }
-
-    // Wait for table to become active.
-    if (!movieTableAlreadyExisted) {
-        std::cout << "Waiting for table '" << MOVIE_TABLE_NAME
-                  << "' to become active...." << std::endl;
-        if (!AwsDoc::DynamoDB::waitTableActive(MOVIE_TABLE_NAME, dynamoClient)) {
-            deleteDynamoTable(MOVIE_TABLE_NAME, dynamoClient);
-            return false;
-        }
-        std::cout << "Table '" << MOVIE_TABLE_NAME << "' created and active."
-                  << std::endl;
-    }
 
     // 2. Add a new movie.
     Aws::String title;
@@ -281,7 +127,6 @@ bool AwsDoc::DynamoDB::dynamodbGettingStartedScenario(
                 putItemRequest);
         if (!outcome.IsSuccess()) {
             std::cerr << "Failed to add an item: " << outcome.GetError().GetMessage() << std::endl;
-            deleteDynamoTable(MOVIE_TABLE_NAME, dynamoClient);
             return false;
         }
     }
@@ -294,7 +139,7 @@ bool AwsDoc::DynamoDB::dynamodbGettingStartedScenario(
         rating = askQuestionForFloatRange(
                 Aws::String("\nLet's update your movie.\nYou rated it  ") +
                 std::to_string(rating)
-                + ", what new would you give it? ", 1, 10);
+                + ", what new rating would you give it? ", 1, 10);
         plot = askQuestion(Aws::String("You summarized the plot as '") + plot +
                            "'.\nWhat would you say now? ");
 
@@ -320,7 +165,6 @@ bool AwsDoc::DynamoDB::dynamodbGettingStartedScenario(
         if (!result.IsSuccess()) {
             std::cerr << "Error updating movie " + result.GetError().GetMessage()
                       << std::endl;
-            deleteDynamoTable(MOVIE_TABLE_NAME, dynamoClient);
             return false;
         }
     }
@@ -328,7 +172,7 @@ bool AwsDoc::DynamoDB::dynamodbGettingStartedScenario(
     std::cout << "\nUpdated '" << title << "' with new attributes:" << std::endl;
 
     // 4. Put 250 movies in the table from moviedata.json.
-    if (!movieTableAlreadyExisted) {
+    {
         std::cout << "Adding movies from a json file to the database." << std::endl;
         const size_t MAX_SIZE_FOR_BATCH_WRITE = 25;
         const size_t MOVIES_TO_WRITE = 10 * MAX_SIZE_FOR_BATCH_WRITE;
@@ -544,59 +388,9 @@ bool AwsDoc::DynamoDB::dynamodbGettingStartedScenario(
         }
     }
 
-    // 9.Delete the table. (DeleteTable)
-    return deleteDynamoTable(MOVIE_TABLE_NAME, dynamoClient);
+    return true;
 }
 
-bool AwsDoc::DynamoDB::deleteDynamoTable(const Aws::String &tableName,
-                                         const Aws::DynamoDB::DynamoDBClient &dynamoClient) {
-    Aws::DynamoDB::Model::DeleteTableRequest request;
-    request.SetTableName(tableName);
-
-    const Aws::DynamoDB::Model::DeleteTableOutcome &result = dynamoClient.DeleteTable(
-            request);
-    if (result.IsSuccess()) {
-        std::cout << "Your Table \""
-                  << result.GetResult().GetTableDescription().GetTableName()
-                  << " was deleted!\n";
-    }
-    else {
-        std::cerr << "Failed to delete table: " << result.GetError().GetMessage() << std::endl;
-    }
-
-    return result.IsSuccess();
-}
-
-bool AwsDoc::DynamoDB::waitTableActive(const Aws::String &tableName,
-                                       const Aws::DynamoDB::DynamoDBClient &dynamoClient) {
-    // Repeatedly call DescribeTable until table is ACTIVE.
-    const int MAX_QUERIES = 20;
-    Aws::DynamoDB::Model::DescribeTableRequest request;
-    request.SetTableName(tableName);
-
-    int count = 0;
-    while (count < MAX_QUERIES) {
-        const Aws::DynamoDB::Model::DescribeTableOutcome &result = dynamoClient.DescribeTable(
-                request);
-        if (result.IsSuccess()) {
-            Aws::DynamoDB::Model::TableStatus status = result.GetResult().GetTable().GetTableStatus();
-
-            if (Aws::DynamoDB::Model::TableStatus::ACTIVE != status) {
-                std::this_thread::sleep_for(std::chrono::seconds(1));
-            }
-            else {
-                return true;
-            }
-        }
-        else {
-            std::cerr << "Error DynamoDB::waitTableActive "
-                      << result.GetError().GetMessage() << std::endl;
-            return false;
-         }
-        count++;
-    }
-    return false;
-}
 
 Aws::Map<Aws::String, Aws::DynamoDB::Model::AttributeValue>
 AwsDoc::DynamoDB::movieJsonViewToAttributeMap(
@@ -639,185 +433,29 @@ int main(int argc, char **argv) {
     Aws::SDKOptions options;
     InitAPI(options);
 
+    // snippet-start:[cpp.example_code.dynamodb.Scenario_GettingStarted.main]
     {
         Aws::Client::ClientConfiguration clientConfig;
-        AwsDoc::DynamoDB::dynamodbGettingStartedScenario(clientConfig);
+       //  1. Create a table with partition: year (N) and sort: title (S). (CreateTable)
+        if (AwsDoc::DynamoDB::createDynamoDBTable(AwsDoc::DynamoDB::MOVIE_TABLE_NAME, clientConfig)) {
+
+            AwsDoc::DynamoDB::dynamodbGettingStartedScenario(clientConfig);
+
+            // 9.Delete the table. (DeleteTable)
+            AwsDoc::DynamoDB::deleteDynamoTable(AwsDoc::DynamoDB::MOVIE_TABLE_NAME, clientConfig);
+        }
     }
+    // snippet-end:[cpp.example_code.dynamodb.Scenario_GettingStarted.main]
 
     return 0;
 }
 
 #endif // TESTING_BUILD
 
-Aws::String AwsDoc::DynamoDB::askQuestion(const Aws::String &string,
-                                          const std::function<bool(
-                                                  Aws::String)> &test) {
-    Aws::String result;
-    do {
-        std::cout << string;
-        std::getline(std::cin, result);
-        if (result.empty()) {
-            std::cout << "Please enter some text." << std::endl;
-        }
-        if (!test(result)) {
-            continue;
-        }
-    } while (result.empty());
-
-    return result;
-}
-
-int AwsDoc::DynamoDB::askQuestionForInt(const Aws::String &string) {
-    Aws::String resultString = askQuestion(string,
-                                           [](const Aws::String &string1) -> bool {
-                                                   try {
-                                                       (void)std::stoi(string1);
-                                                       return true;
-                                                   }
-                                                   catch (const std::invalid_argument &) {
-                                                       return false;
-                                                   }
-                                           });
-
-    int result = 0;
-    try {
-        result = std::stoi(resultString);
-    }
-    catch (const std::invalid_argument &) {
-        std::cerr << "DynamoDB::askQuestionForInt string not an int "
-                  << resultString << std::endl;
-    }
-    return result;
-}
-
-float AwsDoc::DynamoDB::askQuestionForFloatRange(const Aws::String &string, float low,
-                                                 float high) {
-    Aws::String resultString = askQuestion(string, [low, high](
-            const Aws::String &string1) -> bool {
-            try {
-                float number = std::stof(string1);
-                return number >= low && number <= high;
-            }
-            catch (const std::invalid_argument &) {
-                return false;
-            }
-    });
-    float result = 0;
-    try {
-        result = std::stof(resultString);
-    }
-    catch (const std::invalid_argument &) {
-        std::cerr << "DynamoDB::askQuestionForFloatRange string not an int "
-                  << resultString << std::endl;
-    }
-
-    return result;
-}
-
-int AwsDoc::DynamoDB::askQuestionForIntRange(const Aws::String &string, int low,
-                                             int high) {
-    Aws::String resultString = askQuestion(string, [low, high](
-            const Aws::String &string1) -> bool {
-            try {
-                int number = std::stoi(string1);
-                return number >= low && number <= high;
-            }
-            catch (const std::invalid_argument &) {
-                return false;
-            }
-    });
-    int result = 0;
-    try {
-        result = std::stoi(resultString);
-    }
-    catch (const std::invalid_argument &) {
-        std::cerr << "DynamoDB::askQuestionForFloatRange string not an int "
-                  << resultString << std::endl;
-    }
-
-    return result;
-}
-
-void AwsDoc::DynamoDB::printMovieInfo(
-        const Aws::Map<Aws::String, Aws::DynamoDB::Model::AttributeValue> &movieMap) {
-    {
-        auto const &iter = movieMap.find(TITLE_KEY);
-        if (iter != movieMap.end()) {
-            std::cout << "Movie title: '" + iter->second.GetS() << "'." << std::endl;
-        }
-    }
-
-    {
-        auto const &iter = movieMap.find(YEAR_KEY);
-        if (iter != movieMap.end()) {
-            std::cout << "    Year: " + iter->second.GetN() << "." << std::endl;
-        }
-    }
-
-    {
-        auto const &iter = movieMap.find(INFO_KEY);
-        if (iter != movieMap.end()) {
-            Aws::Map<Aws::String, const std::shared_ptr<Aws::DynamoDB::Model::AttributeValue>> infoMap =
-                    iter->second.GetM();
-
-            auto const &ratingIter = infoMap.find(RATING_KEY);
-            if (ratingIter != infoMap.end()) {
-                std::cout << "    Rating: " + ratingIter->second->GetN() << "."
-                          << std::endl;
-            }
-
-            auto const &plotIter = infoMap.find(PLOT_KEY);
-            if (plotIter != infoMap.end()) {
-                std::cout << "    Synopsis: " + plotIter->second->GetS() << "."
-                          << std::endl;
-            }
-        }
-    }
-}
-
-static int deflateZip(FILE *source, FILE *dest);
-
 Aws::String AwsDoc::DynamoDB::getMovieJSON() {
     const int BUFFER_SIZE = 1024;
-    const Aws::String JSON_FILE_NAME("moviedata.json");
     Aws::String result;
-
-#ifdef _HAS_ZLIB_
-    const Aws::String ZIP_FILE_NAME("movie.zip");
-
-    Aws::Client::ClientConfiguration config;
-
-    auto httpClient = Aws::Http::CreateHttpClient(config);
-    auto request = Aws::Http::CreateHttpRequest(Aws::String(
-                                                        "https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/samples/moviedata.zip"),
-                                                Aws::Http::HttpMethod::HTTP_GET,
-                                                Aws::Utils::Stream::DefaultResponseStreamFactoryMethod);
-    request->SetUserAgent("curl/7.79.1");
-    std::cout << "Downloading the json file." << std::endl;
-    auto response = httpClient->MakeRequest(request);
-
-    if (Aws::Http::HttpResponseCode::OK == response->GetResponseCode()) {
-        {
-            std::ofstream outStream(ZIP_FILE_NAME);
-            outStream << response->GetResponseBody().rdbuf();
-        }
-        FILE *src = fopen(ZIP_FILE_NAME.c_str(), "r");
-        FILE *dst = fopen(JSON_FILE_NAME.c_str(), "w");
-
-        std::cout << "Unzipping the json file." << std::endl;
-        int zipResult = deflateZip(src, dst);
-        if (zipResult != Z_OK) {
-            std::cerr << "Could not deflate zip file" << std::endl;
-        }
-        fclose(src);
-        fclose(dst);
-    }
-    else {
-        std::cerr << "Could not download json File "
-                  << response->GetClientErrorMessage() << std::endl;
-    }
-#endif //_HAS_ZLIB_
-    std::ifstream movieData(JSON_FILE_NAME);
+    std::ifstream movieData(MOVIE_FILE_PATH);  // MOVIE_FILE_PATH is defined in CMakeLists.txt.
     if (movieData) { // NOLINT (readability-implicit-bool-conversion)
         std::array<char, BUFFER_SIZE> buffer{};
         while (movieData) { // NOLINT (readability-implicit-bool-conversion)
@@ -826,90 +464,5 @@ Aws::String AwsDoc::DynamoDB::getMovieJSON() {
             result += &buffer[0];
         }
     }
-    else {
-        std::cerr << "Could not open '" << JSON_FILE_NAME << "'." << std::endl;
-#ifndef _HAS_ZLIB_
-        std::cerr << "This app was built without zlib." << std::endl;
-        std::cerr << "To run the complete scenario, install zlib or" << std::endl;
-        std::cerr << "download and unzip the following file to your run directory."
-                  << std::endl;
-        std::cerr
-                << "https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/samples/moviedata.zip"
-                << std::endl;
-#endif //_HAS_ZLIB_
-    }
-
     return result;
 }
-
-#ifdef _HAS_ZLIB_
-int deflateZip(FILE *source, FILE *dest) {
-    const int IN_CHUNK = 32767;
-    const int OUT_CHUNK = 65535;
-    int ret;
-    unsigned have;
-    z_stream strm = {};
-    unsigned char in[IN_CHUNK];
-    unsigned char out[OUT_CHUNK];
-
-    // Read to the end of the local file header.
-    struct __attribute__((__packed__)) ZipHeader {
-        uint16_t ignored[13];
-        uint16_t fileNameLength;
-        uint16_t extraFieldLength;
-    };
-    ZipHeader header{};
-
-    fread(&header, 1, sizeof(header), source);
-    fread(in, 1, header.fileNameLength + header.extraFieldLength, source);
-    // Local file header read.
-
-    strm.data_type = Z_BINARY;
-    strm.zalloc = Z_NULL;
-    strm.zfree = Z_NULL;
-    strm.opaque = Z_NULL;
-    strm.avail_in = 0;
-    strm.next_in = Z_NULL;
-    int windowBits = -MAX_WBITS;
-    ret = inflateInit2(&strm, windowBits);
-    if (ret != Z_OK)
-        return ret;
-
-     do {
-        strm.avail_in = fread(in, 1, IN_CHUNK, source);
-        if (ferror(source)) {
-            (void) inflateEnd(&strm);
-            return Z_ERRNO;
-        }
-        if (strm.avail_in == 0)
-            break;
-        strm.next_in = in;
-
-         do {
-            strm.avail_out = OUT_CHUNK;
-            strm.next_out = out;
-            ret = inflate(&strm, Z_SYNC_FLUSH);
-            assert(ret != Z_STREAM_ERROR);  // State not clobbered.
-            switch (ret) {
-                case Z_NEED_DICT:
-                    ret = Z_DATA_ERROR;     // And fall through.
-                case Z_DATA_ERROR:
-                case Z_MEM_ERROR:
-                    (void) inflateEnd(&strm);
-                    return ret;
-                default:
-                    break;
-            }
-            have = OUT_CHUNK - strm.avail_out;
-            if (fwrite(out, 1, have, dest) != have || ferror(dest)) {
-                (void) inflateEnd(&strm);
-                return Z_ERRNO;
-            }
-        } while (strm.avail_out == 0);
-
-     } while (ret != Z_STREAM_END);
-
-    (void) inflateEnd(&strm);
-    return ret == Z_STREAM_END ? Z_OK : Z_DATA_ERROR;
-}
-#endif
