+++ cpp/example_code/dynamodb/dynamodb_partiql_scenario.cpp
@@ -33,9 +33,6 @@
 #include <iomanip>
 #include <aws/core/Aws.h>
 #include <aws/dynamodb/DynamoDBClient.h>
-#include <aws/dynamodb/model/CreateTableRequest.h>
-#include <aws/dynamodb/model/DeleteTableRequest.h>
-#include <aws/dynamodb/model/DescribeTableRequest.h>
 #include <aws/dynamodb/model/ExecuteStatementRequest.h>
 #include <aws/dynamodb/model/BatchExecuteStatementRequest.h>
 
@@ -43,174 +40,15 @@
 
 namespace AwsDoc {
     namespace DynamoDB {
-        /**
-         * Constants used for DynamoDB table creation.
-         */
-        static const Aws::String MOVIE_TABLE_NAME("doc-example-table-movies");
-        static const Aws::String YEAR_KEY("year");
-        static const Aws::String TITLE_KEY("title");
-        static const Aws::String INFO_KEY("info");
-        static const Aws::String RATING_KEY("rating");
-        static const Aws::String PLOT_KEY("plot");
-        static const int PROVISIONED_THROUGHPUT_UNITS = 10;
-        static const Aws::String ALLOCATION_TAG("dynamodb_scenario");
-        static const int ASTERIX_FILL_WIDTH = 88;
 
-        //! Delete a DynamoDB table.
-        /*!
-          \sa deleteDynamoTable()
-          \param tableName: The DynamoDB table's name.
-          \param dynamoClient: A DynamoDB client.
-          \return bool: Function succeeded.
-        */
-        static bool deleteDynamoTable(const Aws::String &tableName,
-                                      const Aws::DynamoDB::DynamoDBClient &dynamoClient);
-
-        //! Query a newly created DynamoDB table until it is active.
-        /*!
-          \sa waitTableActive()
-          \param waitTableActive: The DynamoDB table's name.
-          \param dynamoClient: A DynamoDB client.
-          \return bool: Function succeeded.
-        */
-        static bool waitTableActive(const Aws::String &tableName,
-                                    const Aws::DynamoDB::DynamoDBClient &dynamoClient);
-
-        //! Command line prompt/response utility function.
-        /*!
-         \\sa askQuestion()
-         \param string: A question prompt.
-         \param test: Test function for response.
-         \return Aws::String: User's response.
-         */
-        static Aws::String askQuestion(const Aws::String &string,
-                                       const std::function<bool(
-                                               Aws::String)> &test = [](
-                                               const Aws::String &) -> bool { return true; });
-
-        //! Command line prompt/response utility function for an integer result.
-        /*!
-         \sa askQuestionForInt()
-         \param string: A question prompt.
-         \return int: User's response.
-         */
-        int askQuestionForInt(const std::string &string);
-
-        //! Command line prompt/response utility function for a float result confined to
-        //! a range.
-        /*!
-         \sa askQuestionForFloatRange()
-         \param string: A question prompt.
-         \param low: Low inclusive.
-         \param high: High inclusive.
-         \return float: User's response.
-         */
-        float
-        askQuestionForFloatRange(const Aws::String &string, float low, float high);
-
-        //! Command line prompt/response utility function for an int result confined to
-        //! a range.
-        /*!
-         \sa askQuestionForIntRange()
-         \param string: A question prompt.
-         \param low: Low inclusive.
-         \param high: High inclusive.
-         \return int: User's response.
-         */
-        int askQuestionForIntRange(const Aws::String &string, int low,
-                                   int high);
-
-        //! Utility function to log movie attributes to std::cout.
-        /*!
-         \sa printMovieInfo()
-         \param movieMap: Map of DynamoDB attribute values.
-         \return void
-         */
-        void printMovieInfo(
-                const Aws::Map<Aws::String, Aws::DynamoDB::Model::AttributeValue> &movieMap);
+        static bool dynamodbPartiqlScenario(const Aws::Client::ClientConfiguration& clientConfiguration);
 
      } //  namespace DynamoDB
 } // namespace AwsDoc
 
-//! Scenario to create, modify, query, and delete a DynamoDB table.
-/*!
-  \sa dynamodbGettingStartedScenario()
-  \param clientConfig: Aws client configuration.
-  \return bool: Function succeeded.
- */
-
-// snippet-start:[cpp.example_code.dynamodb.Scenario_GettingStarted]
-bool AwsDoc::DynamoDB::dynamodbGettingStartedScenario(
-        const Aws::Client::ClientConfiguration &clientConfiguration) {
-    std::cout << std::setfill('*') << std::setw(ASTERIX_FILL_WIDTH) << " " << std::endl;
-    std::cout << "Welcome to the Amazon DynamoDB getting started demo." << std::endl;
-    std::cout << std::setfill('*') << std::setw(ASTERIX_FILL_WIDTH) << " " << std::endl;
-
+bool
+AwsDoc::DynamoDB::partiqlExecuteScenario(const Aws::Client::ClientConfiguration& clientConfiguration) {
     Aws::DynamoDB::DynamoDBClient dynamoClient(clientConfiguration);
-
-    bool movieTableAlreadyExisted = false;
-
-    // 1. Create a table.
-    {
-        Aws::DynamoDB::Model::CreateTableRequest request;
-
-        Aws::DynamoDB::Model::AttributeDefinition yearAttributeDefinition;
-        yearAttributeDefinition.SetAttributeName(YEAR_KEY);
-        yearAttributeDefinition.SetAttributeType(
-                Aws::DynamoDB::Model::ScalarAttributeType::N);
-        request.AddAttributeDefinitions(yearAttributeDefinition);
-
-        Aws::DynamoDB::Model::AttributeDefinition titleAttributeDefinition;
-        yearAttributeDefinition.SetAttributeName(TITLE_KEY);
-        yearAttributeDefinition.SetAttributeType(
-                Aws::DynamoDB::Model::ScalarAttributeType::S);
-        request.AddAttributeDefinitions(yearAttributeDefinition);
-
-        Aws::DynamoDB::Model::KeySchemaElement yearKeySchema;
-        yearKeySchema.WithAttributeName(YEAR_KEY).WithKeyType(
-                Aws::DynamoDB::Model::KeyType::HASH);
-        request.AddKeySchema(yearKeySchema);
-
-        Aws::DynamoDB::Model::KeySchemaElement titleKeySchema;
-        yearKeySchema.WithAttributeName(TITLE_KEY).WithKeyType(
-                Aws::DynamoDB::Model::KeyType::RANGE);
-        request.AddKeySchema(yearKeySchema);
-
-        Aws::DynamoDB::Model::ProvisionedThroughput throughput;
-        throughput.WithReadCapacityUnits(
-                PROVISIONED_THROUGHPUT_UNITS).WithWriteCapacityUnits(
-                PROVISIONED_THROUGHPUT_UNITS);
-        request.SetProvisionedThroughput(throughput);
-        request.SetTableName(MOVIE_TABLE_NAME);
-
-        std::cout << "Creating table '" << MOVIE_TABLE_NAME << "'..." << std::endl;
-        const Aws::DynamoDB::Model::CreateTableOutcome &result = dynamoClient.CreateTable(
-                request);
-        if (!result.IsSuccess()) {
-            if (result.GetError().GetErrorType() ==
-                Aws::DynamoDB::DynamoDBErrors::RESOURCE_IN_USE) {
-                std::cout << "Table already exists." << std::endl;
-                movieTableAlreadyExisted = true;
-            }
-            else {
-                std::cerr << "Failed to create table: "
-                          << result.GetError().GetMessage();
-                return false;
-            }
-        }
-    }
-
-    // Wait for table to become active.
-    if (!movieTableAlreadyExisted) {
-        std::cout << "Waiting for table '" << MOVIE_TABLE_NAME
-                  << "' to become active...." << std::endl;
-        if (!AwsDoc::DynamoDB::waitTableActive(MOVIE_TABLE_NAME, dynamoClient)) {
-            deleteDynamoTable(MOVIE_TABLE_NAME, dynamoClient);
-            return false;
-        }
-        std::cout << "Table '" << MOVIE_TABLE_NAME << "' created and active."
-                  << std::endl;
-    }
 
     // 2. Add a new movie using an "Insert" statement. (ExecuteStatement)
     Aws::String title;
@@ -228,9 +66,9 @@ bool AwsDoc::DynamoDB::dynamodbGettingStartedScenario(
         Aws::DynamoDB::Model::ExecuteStatementRequest request;
         std::stringstream sqlStream;
         sqlStream << "INSERT INTO \"" << MOVIE_TABLE_NAME << "\" VALUE {'"
-                << TITLE_KEY << "': ?, '" << YEAR_KEY << "': ?, '"
-                << INFO_KEY << "': ?}";
-        std::cout << "SQL statement " << sqlStream.str() << std::endl;
+                  << TITLE_KEY << "': ?, '" << YEAR_KEY << "': ?, '"
+                  << INFO_KEY << "': ?}";
+
         request.SetStatement(sqlStream.str());
 
         Aws::Vector<Aws::DynamoDB::Model::AttributeValue> attributes;
@@ -257,7 +95,6 @@ bool AwsDoc::DynamoDB::dynamodbGettingStartedScenario(
                 request);
         if (!outcome.IsSuccess()) {
             std::cerr << "Failed to add an item: " << outcome.GetError().GetMessage() << std::endl;
-            deleteDynamoTable(MOVIE_TABLE_NAME, dynamoClient);
             return false;
         }
     }
@@ -265,14 +102,14 @@ bool AwsDoc::DynamoDB::dynamodbGettingStartedScenario(
     std::cout << "\nAdded '" << title << "' to '" << MOVIE_TABLE_NAME << "'."
               << std::endl;
 
-  //  3. Get the data for the movie using a "Select" statement. (ExecuteStatement)
+    //  3. Get the data for the movie using a "Select" statement. (ExecuteStatement)
     {
         Aws::DynamoDB::Model::ExecuteStatementRequest request;
         std::stringstream sqlStream;
         sqlStream << "SELECT * FROM  \"" << MOVIE_TABLE_NAME << "\" WHERE "
                   << TITLE_KEY << "=? and " << YEAR_KEY << "=?";
-        std::cout << "SQL statement " << sqlStream.str() << std::endl;
-        request.SetStatement(sqlStream.str());
+
+         request.SetStatement(sqlStream.str());
 
         Aws::Vector<Aws::DynamoDB::Model::AttributeValue> attributes;
         attributes.push_back(Aws::DynamoDB::Model::AttributeValue().SetS(title));
@@ -283,7 +120,6 @@ bool AwsDoc::DynamoDB::dynamodbGettingStartedScenario(
                 request);
         if (!outcome.IsSuccess()) {
             std::cerr << "Failed to retrieve item information: " << outcome.GetError().GetMessage() << std::endl;
-            deleteDynamoTable(MOVIE_TABLE_NAME, dynamoClient);
             return false;
         }
         else{
@@ -297,24 +133,24 @@ bool AwsDoc::DynamoDB::dynamodbGettingStartedScenario(
             }
             else{
                 std::cerr << "Error: " << items.size() << " items were retrieved. "
-                << " There should be only one item." << std::endl;
+                          << " There should be only one item." << std::endl;
             }
         }
     }
 
-  //  4. Update the data for the movie using an "Update" statement. (ExecuteStatement)
+    //  4. Update the data for the movie using an "Update" statement. (ExecuteStatement)
     {
         rating = askQuestionForFloatRange(
                 Aws::String("\nLet's update your movie.\nYou rated it  ") +
                 std::to_string(rating)
-                + ", what new would you give it? ", 1, 10);
+                + ", what new rating would you give it? ", 1, 10);
 
         Aws::DynamoDB::Model::ExecuteStatementRequest request;
         std::stringstream sqlStream;
         sqlStream << "UPDATE \"" << MOVIE_TABLE_NAME << "\" SET "
                   << INFO_KEY << "." << RATING_KEY << "=? WHERE "
                   << TITLE_KEY << "=? AND " << YEAR_KEY << "=?";
-        std::cout << "SQL statement " << sqlStream.str() << std::endl;
+
         request.SetStatement(sqlStream.str());
 
         Aws::Vector<Aws::DynamoDB::Model::AttributeValue> attributes;
@@ -328,10 +164,9 @@ bool AwsDoc::DynamoDB::dynamodbGettingStartedScenario(
                 request);
         if (!outcome.IsSuccess()) {
             std::cerr << "Failed to update an item: " << outcome.GetError().GetMessage();
-            deleteDynamoTable(MOVIE_TABLE_NAME, dynamoClient);
             return false;
         }
-     }
+    }
 
     std::cout << "\nUpdated '" << title << "' with new attributes:" << std::endl;
 
@@ -341,7 +176,7 @@ bool AwsDoc::DynamoDB::dynamodbGettingStartedScenario(
         std::stringstream sqlStream;
         sqlStream << "SELECT * FROM  \"" << MOVIE_TABLE_NAME << "\" WHERE "
                   << TITLE_KEY << "=? and " << YEAR_KEY << "=?";
-        std::cout << "SQL statement " << sqlStream.str() << std::endl;
+
         request.SetStatement(sqlStream.str());
 
         Aws::Vector<Aws::DynamoDB::Model::AttributeValue> attributes;
@@ -353,7 +188,6 @@ bool AwsDoc::DynamoDB::dynamodbGettingStartedScenario(
                 request);
         if (!outcome.IsSuccess()) {
             std::cerr << "Failed to retrieve item information: " << outcome.GetError().GetMessage() << std::endl;
-            deleteDynamoTable(MOVIE_TABLE_NAME, dynamoClient);
             return false;
         }
         else{
@@ -380,8 +214,8 @@ bool AwsDoc::DynamoDB::dynamodbGettingStartedScenario(
         std::stringstream sqlStream;
         sqlStream << "DELETE FROM  \"" << MOVIE_TABLE_NAME << "\" WHERE "
                   << TITLE_KEY << "=? and " << YEAR_KEY << "=?";
-        std::cout << "SQL statement " << sqlStream.str() << std::endl;
-        request.SetStatement(sqlStream.str());
+
+         request.SetStatement(sqlStream.str());
 
         Aws::Vector<Aws::DynamoDB::Model::AttributeValue> attributes;
         attributes.push_back(Aws::DynamoDB::Model::AttributeValue().SetS(title));
@@ -392,14 +226,19 @@ bool AwsDoc::DynamoDB::dynamodbGettingStartedScenario(
                 request);
         if (!outcome.IsSuccess()) {
             std::cerr << "Failed to delete the move: " << outcome.GetError().GetMessage() << std::endl;
-            deleteDynamoTable(MOVIE_TABLE_NAME, dynamoClient);
             return false;
         }
     }
 
     std::cout << "Movie successfully deleted." << std::endl;
+    return true;
+}
 
+bool AwsDoc::DynamoDB::partiqlBatchExecuteScenario(
+        const Aws::Client::ClientConfiguration& clientConfiguration) {
     std::cout << "Now we will work with batches of movies" << std::endl;
+
+    Aws::DynamoDB::DynamoDBClient dynamoClient(clientConfiguration);
 
 
     std::vector<Aws::String> titles;
@@ -414,287 +253,257 @@ bool AwsDoc::DynamoDB::dynamodbGettingStartedScenario(
         int aYear = askQuestionForInt("What year was it released? ");
         years.push_back(aYear);
         float aRating = askQuestionForFloatRange("On a scale of 1 - 10, how do you rate it? ",
-                                          1, 10);
+                                                 1, 10);
         ratings.push_back(aRating);
         Aws::String aPlot = askQuestion("Summarize the plot for me: ");
         plots.push_back(aPlot);
 
-        doAgain = askQuestion(Aws::String("Try another year? (y/n) "));
+        doAgain = askQuestion(Aws::String("Would you like to add more movies? (y/n) "));
     } while (doAgain == "y");
 
-#if 0
-    // 4. Put 250 movies in the table from moviedata.json.
-    if (!movieTableAlreadyExisted) {
-        std::cout << "Adding movies from a json file to the database." << std::endl;
-        const size_t MAX_SIZE_FOR_BATCH_WRITE = 25;
-        const size_t MOVIES_TO_WRITE = 10 * MAX_SIZE_FOR_BATCH_WRITE;
-        Aws::String jsonString = getMovieJSON();
-        if (!jsonString.empty()) {
-            Aws::Utils::Json::JsonValue json(jsonString);
-            Aws::Utils::Array<Aws::Utils::Json::JsonView> movieJsons = json.View().AsArray();
-            Aws::Vector<Aws::DynamoDB::Model::WriteRequest> writeRequests;
+    std::cout << "Adding "  << titles.size() << (titles.size() == 1 ? " movie " : " movies ")
+              << "to the table using a batch \"INSERT\" statement." <<  std::endl;
 
-            // To add movies with a cross-section of years, use an appropriate increment
-            // value for iterating through the database.
-            size_t increment = movieJsons.GetLength() / MOVIES_TO_WRITE;
-            for (size_t i = 0; i < movieJsons.GetLength(); i += increment) {
-                writeRequests.push_back(Aws::DynamoDB::Model::WriteRequest());
-                Aws::Map<Aws::String, Aws::DynamoDB::Model::AttributeValue> putItems = movieJsonViewToAttributeMap(
-                        movieJsons[i]);
-                Aws::DynamoDB::Model::PutRequest putRequest;
-                putRequest.SetItem(putItems);
-                writeRequests.back().SetPutRequest(putRequest);
-                if (writeRequests.size() == MAX_SIZE_FOR_BATCH_WRITE) {
-                    Aws::DynamoDB::Model::BatchWriteItemRequest request;
-                    request.AddRequestItems(MOVIE_TABLE_NAME, writeRequests);
-                    const Aws::DynamoDB::Model::BatchWriteItemOutcome &outcome = dynamoClient.BatchWriteItem(
-                            request);
-                    if (!outcome.IsSuccess()) {
-                        std::cerr << "Unable to batch write movie data: "
-                                  << outcome.GetError().GetMessage()
-                                  << std::endl;
-                        writeRequests.clear();
-                        break;
-                    }
-                    else {
-                        std::cout << "Added batch of " << writeRequests.size()
-                                  << " movies to the database."
-                                  << std::endl;
-                    }
-                    writeRequests.clear();
-                }
-            }
+    {
+        Aws::Vector<Aws::DynamoDB::Model::BatchStatementRequest> statements(titles.size());
+
+        std::stringstream sqlStream;
+        sqlStream << "INSERT INTO \"" << MOVIE_TABLE_NAME << "\" VALUE {'"
+                  << TITLE_KEY << "': ?, '" << YEAR_KEY << "': ?, '"
+                  << INFO_KEY << "': ?}";
+
+        std::string sql(sqlStream.str());
+
+        for (size_t i = 0; i < statements.size(); ++i) {
+            statements[i].SetStatement(sql);
+
+            Aws::Vector<Aws::DynamoDB::Model::AttributeValue> attributes;
+            attributes.push_back(Aws::DynamoDB::Model::AttributeValue().SetS(titles[i]));
+            attributes.push_back(Aws::DynamoDB::Model::AttributeValue().SetN(years[i]));
+
+
+            // Create attribute for the info map.
+            Aws::DynamoDB::Model::AttributeValue infoMapAttribute;
+
+            std::shared_ptr<Aws::DynamoDB::Model::AttributeValue> ratingAttribute = Aws::MakeShared<Aws::DynamoDB::Model::AttributeValue>(
+                    ALLOCATION_TAG.c_str());
+            ratingAttribute->SetN(ratings[i]);
+            infoMapAttribute.AddMEntry(RATING_KEY, ratingAttribute);
+
+            std::shared_ptr<Aws::DynamoDB::Model::AttributeValue> plotAttibute = Aws::MakeShared<Aws::DynamoDB::Model::AttributeValue>(
+                    ALLOCATION_TAG.c_str());
+            plotAttibute->SetS(plots[i]);
+            infoMapAttribute.AddMEntry(PLOT_KEY, plotAttibute);
+            attributes.push_back(infoMapAttribute);
+            statements[i].SetParameters(attributes);
+        }
+
+        Aws::DynamoDB::Model::BatchExecuteStatementRequest request;
+
+        request.SetStatements(statements);
+
+
+        Aws::DynamoDB::Model::BatchExecuteStatementOutcome outcome = dynamoClient.BatchExecuteStatement(
+                request);
+        if (!outcome.IsSuccess()) {
+            std::cerr << "Failed to add an item: " << outcome.GetError().GetMessage() << std::endl;
+            return false;
         }
     }
 
+    std::cout << "Retrieving the movie data with a batch \"SELECT\" statement." << std::endl;
+
+    // 8. Get the data for multiple using a "Select" statement. (BatchExecuteStatement)
+    {
+        Aws::Vector<Aws::DynamoDB::Model::BatchStatementRequest> statements(titles.size());
+        std::stringstream sqlStream;
+        sqlStream << "SELECT * FROM  \"" << MOVIE_TABLE_NAME << "\" WHERE "
+                  << TITLE_KEY << "=? and " << YEAR_KEY << "=?";
+
+        std::string sql(sqlStream.str());
+
+        for (size_t i = 0; i < statements.size(); ++i) {
+            statements[i].SetStatement(sql);
+            Aws::Vector<Aws::DynamoDB::Model::AttributeValue> attributes;
+            attributes.push_back(Aws::DynamoDB::Model::AttributeValue().SetS(titles[i]));
+            attributes.push_back(Aws::DynamoDB::Model::AttributeValue().SetN(years[i]));
+            statements[i].SetParameters(attributes);
+        }
+
+        Aws::DynamoDB::Model::BatchExecuteStatementRequest request;
+
+        request.SetStatements(statements);
+
+        Aws::DynamoDB::Model::BatchExecuteStatementOutcome outcome = dynamoClient.BatchExecuteStatement(
+                request);
+        if (outcome.IsSuccess()) {
+            const Aws::DynamoDB::Model::BatchExecuteStatementResult& result = outcome.GetResult();
+
+            const Aws::Vector<Aws::DynamoDB::Model::BatchStatementResponse>& responses = result.GetResponses();
+
+            for (const Aws::DynamoDB::Model::BatchStatementResponse& response : responses)
+            {
+                const Aws::Map<Aws::String, Aws::DynamoDB::Model::AttributeValue>& item = response.GetItem();
+
+                printMovieInfo(item);
+            }
+        }
+        else{
+             std::cerr << "Failed to retrieve item information: " << outcome.GetError().GetMessage() << std::endl;
+            return false;
+        }
+    }
+
+  //   9. Update the data for multiple using an "Update" statement. (BatchExecuteStatement)
+
+  for (size_t i = 0; i < titles.size(); ++i)
+    {
+        ratings[i] = askQuestionForFloatRange(
+                Aws::String("\nLet's update your the movie, \"") + titles[i] +
+                ".\nYou rated it  " + std::to_string(ratings[i])
+                + ", what new rating would you give it? ", 1, 10);
+    }
+
+    std::cout << "Updating the movie with batch \"UPDATE\" statement." << std::endl;
+
+    {
+        Aws::Vector<Aws::DynamoDB::Model::BatchStatementRequest> statements(titles.size());
+
+        std::stringstream sqlStream;
+        sqlStream << "UPDATE \"" << MOVIE_TABLE_NAME << "\" SET "
+                  << INFO_KEY << "." << RATING_KEY << "=? WHERE "
+                  << TITLE_KEY << "=? AND " << YEAR_KEY << "=?";
+
+
+        std::string sql(sqlStream.str());
+
+        for (size_t i = 0; i < statements.size(); ++i) {
+            statements[i].SetStatement(sql);
+
+            Aws::Vector<Aws::DynamoDB::Model::AttributeValue> attributes;
+            attributes.push_back(Aws::DynamoDB::Model::AttributeValue().SetN(ratings[i]));
+            attributes.push_back(Aws::DynamoDB::Model::AttributeValue().SetS(titles[i]));
+            attributes.push_back(Aws::DynamoDB::Model::AttributeValue().SetN(years[i]));
+            statements[i].SetParameters(attributes);
+        }
+
+        Aws::DynamoDB::Model::BatchExecuteStatementRequest request;
+
+        request.SetStatements(statements);
+        Aws::DynamoDB::Model::BatchExecuteStatementOutcome outcome = dynamoClient.BatchExecuteStatement(
+                request);
+        if (!outcome.IsSuccess()) {
+            std::cerr << "Failed to update movie information: " << outcome.GetError().GetMessage() << std::endl;
+            return false;
+        }
+    }
+
+    std::cout << "Retrieving the updated movie data with a batch \"SELECT\" statement." << std::endl;
+
+    // 10. Get the updated data for multiple movies using a "Select" statement. (BatchExecuteStatement)
+    {
+        Aws::Vector<Aws::DynamoDB::Model::BatchStatementRequest> statements(titles.size());
+        std::stringstream sqlStream;
+        sqlStream << "SELECT * FROM  \"" << MOVIE_TABLE_NAME << "\" WHERE "
+                  << TITLE_KEY << "=? and " << YEAR_KEY << "=?";
+
+        std::string sql(sqlStream.str());
+
+        for (size_t i = 0; i < statements.size(); ++i) {
+            statements[i].SetStatement(sql);
+            Aws::Vector<Aws::DynamoDB::Model::AttributeValue> attributes;
+            attributes.push_back(Aws::DynamoDB::Model::AttributeValue().SetS(titles[i]));
+            attributes.push_back(Aws::DynamoDB::Model::AttributeValue().SetN(years[i]));
+            statements[i].SetParameters(attributes);
+        }
+
+        Aws::DynamoDB::Model::BatchExecuteStatementRequest request;
+
+        request.SetStatements(statements);
+
+        Aws::DynamoDB::Model::BatchExecuteStatementOutcome outcome = dynamoClient.BatchExecuteStatement(
+                request);
+        if (outcome.IsSuccess()) {
+            const Aws::DynamoDB::Model::BatchExecuteStatementResult& result = outcome.GetResult();
+
+            const Aws::Vector<Aws::DynamoDB::Model::BatchStatementResponse>& responses = result.GetResponses();
+
+            for (const Aws::DynamoDB::Model::BatchStatementResponse& response : responses)
+            {
+                const Aws::Map<Aws::String, Aws::DynamoDB::Model::AttributeValue>& item = response.GetItem();
+
+                printMovieInfo(item);
+            }
+        }
+        else{
+            std::cerr << "Failed to retrieve item information: " << outcome.GetError().GetMessage() << std::endl;
+            return false;
+        }
+    }
+
+    // 11. Delete the multiple movies using a "Delete" statement. (BatchExecuteStatement)
+
+    std::cout << "Deleting the movie data with a batch \"DELETE\" statement." << std::endl;
+
+    {
+        Aws::Vector<Aws::DynamoDB::Model::BatchStatementRequest> statements(titles.size());
+        std::stringstream sqlStream;
+        sqlStream << "DELETE FROM  \"" << MOVIE_TABLE_NAME << "\" WHERE "
+                  << TITLE_KEY << "=? and " << YEAR_KEY << "=?";
+
+        std::string sql(sqlStream.str());
+
+        for (size_t i = 0; i < statements.size(); ++i) {
+            statements[i].SetStatement(sql);
+            Aws::Vector<Aws::DynamoDB::Model::AttributeValue> attributes;
+            attributes.push_back(Aws::DynamoDB::Model::AttributeValue().SetS(titles[i]));
+            attributes.push_back(Aws::DynamoDB::Model::AttributeValue().SetN(years[i]));
+            statements[i].SetParameters(attributes);
+        }
+
+        Aws::DynamoDB::Model::BatchExecuteStatementRequest request;
+
+        request.SetStatements(statements);
+
+        Aws::DynamoDB::Model::BatchExecuteStatementOutcome outcome = dynamoClient.BatchExecuteStatement(
+                request);
+        if (!outcome.IsSuccess()) {
+            std::cerr << "Failed to delete the movies: " << outcome.GetError().GetMessage() << std::endl;
+            return false;
+        }
+    }
+
+    return true;
+}
+
+
+//! Scenario to create, modify, query, and delete a DynamoDB table.
+/*!
+  \sa dynamodbPartiqlScenario()
+  \param clientConfig: Aws client configuration.
+  \return bool: Function succeeded.
+ */
+
+// snippet-start:[cpp.example_code.dynamodb.Scenario_PartiQL]
+bool AwsDoc::DynamoDB::dynamodbPartiqlScenario(
+        const Aws::Client::ClientConfiguration &clientConfiguration) {
+    std::cout << std::setfill('*') << std::setw(ASTERIX_FILL_WIDTH) << " " << std::endl;
+    std::cout << "Welcome to the Amazon DynamoDB PartiQL demo." << std::endl;
     std::cout << std::setfill('*') << std::setw(ASTERIX_FILL_WIDTH) << " " << std::endl;
 
-    // 5. Get a movie by Key (partition + sort).
-    {
-        Aws::String titleToGet("King Kong");
-        Aws::String answer = askQuestion(Aws::String(
-                "Let's move on...Would you like to get info about '" + titleToGet +
-                "'? (y/n) "));
-        if (answer == "y") {
-            Aws::DynamoDB::Model::GetItemRequest request;
-            request.SetTableName(MOVIE_TABLE_NAME);
-            request.AddKey(TITLE_KEY,
-                           Aws::DynamoDB::Model::AttributeValue().SetS(titleToGet));
-            request.AddKey(YEAR_KEY, Aws::DynamoDB::Model::AttributeValue().SetN(1933));
 
-            const Aws::DynamoDB::Model::GetItemOutcome &result = dynamoClient.GetItem(
-                    request);
-            if (!result.IsSuccess()) {
-                std::cerr << "Error " << result.GetError().GetMessage();
-             }
-            else {
-                const Aws::Map<Aws::String, Aws::DynamoDB::Model::AttributeValue> &item = result.GetResult().GetItem();
-                if (!item.empty()) {
-                    std::cout << "\nHere's what I found:" << std::endl;
-                    printMovieInfo(item);
-                }
-                else {
-                    std::cout << "\nThe movie was not found in the database."
-                              << std::endl;
-                }
-            }
-        }
+
+    bool result = partiqlExecuteScenario(clientConfiguration);
+
+    if (result)
+    {
+        result = partiqlBatchExecuteScenario(clientConfiguration);
     }
 
-    // 6. Use Query with a key condition expression to return all movies
-    //    released in a given year.
-    Aws::String doAgain = "n";
-    do {
-        Aws::DynamoDB::Model::QueryRequest req;
-
-        req.SetTableName(MOVIE_TABLE_NAME);
-
-        // "year" is a DynamoDB reserved keyword and must be replaced with an
-        // expression attribute name.
-        req.SetKeyConditionExpression("#dynobase_year = :valueToMatch");
-        req.SetExpressionAttributeNames({{"#dynobase_year", YEAR_KEY}});
-
-        int yearToMatch = askQuestionForIntRange(
-                "\nLet's get a list of movies released in"
-                " a given year. Enter a year between 1972 and 2018 ",
-                1972, 2018);
-        Aws::Map<Aws::String, Aws::DynamoDB::Model::AttributeValue> attributeValues;
-        attributeValues.emplace(":valueToMatch",
-                                Aws::DynamoDB::Model::AttributeValue().SetN(
-                                        yearToMatch));
-        req.SetExpressionAttributeValues(attributeValues);
-
-        const Aws::DynamoDB::Model::QueryOutcome &result = dynamoClient.Query(req);
-        if (result.IsSuccess()) {
-            const Aws::Vector<Aws::Map<Aws::String, Aws::DynamoDB::Model::AttributeValue>> &items = result.GetResult().GetItems();
-            if (!items.empty()) {
-                std::cout << "\nThere were " << items.size()
-                          << " movies in the database from "
-                          << yearToMatch << "." << std::endl;
-                for (const auto &item: items) {
-                    printMovieInfo(item);
-                }
-                doAgain = "n";
-            }
-            else {
-                std::cout << "\nNo movies from " << yearToMatch
-                          << " were found in the database"
-                          << std::endl;
-                doAgain = askQuestion(Aws::String("Try another year? (y/n) "));
-            }
-        }
-        else {
-            std::cerr << "Failed to Query items: " << result.GetError().GetMessage();
-        }
-
-    } while (doAgain == "y");
-
-    //  7. Use Scan to return movies released within a range of years.
-    //     Show how to paginate data using ExclusiveStartKey. (Scan + FilterExpression)
-    {
-        int startYear = askQuestionForIntRange("\nNow let's scan a range of years "
-                                               "for movies in the database. Enter a start year: ",
-                                               1972, 2018);
-        int endYear = askQuestionForIntRange("\nEnter an end year: ",
-                                             startYear, 2018);
-        Aws::Map<Aws::String, Aws::DynamoDB::Model::AttributeValue> exclusiveStartKey;
-        do {
-            Aws::DynamoDB::Model::ScanRequest scanRequest;
-            scanRequest.SetTableName(MOVIE_TABLE_NAME);
-            scanRequest.SetFilterExpression(
-                    "#dynobase_year >= :startYear AND #dynobase_year <= :endYear");
-            scanRequest.SetExpressionAttributeNames({{"#dynobase_year", YEAR_KEY}});
-
-            Aws::Map<Aws::String, Aws::DynamoDB::Model::AttributeValue> attributeValues;
-            attributeValues.emplace(":startYear",
-                                    Aws::DynamoDB::Model::AttributeValue().SetN(
-                                            startYear));
-            attributeValues.emplace(":endYear",
-                                    Aws::DynamoDB::Model::AttributeValue().SetN(
-                                            endYear));
-            scanRequest.SetExpressionAttributeValues(attributeValues);
-
-            if (!exclusiveStartKey.empty()) {
-                scanRequest.SetExclusiveStartKey(exclusiveStartKey);
-            }
-
-            const Aws::DynamoDB::Model::ScanOutcome &result = dynamoClient.Scan(
-                    scanRequest);
-            if (result.IsSuccess()) {
-                const Aws::Vector<Aws::Map<Aws::String, Aws::DynamoDB::Model::AttributeValue>> &items = result.GetResult().GetItems();
-                if (!items.empty()) {
-                    std::stringstream stringStream;
-                    stringStream << "\nFound " << items.size() << " movies in one scan."
-                                 << " How many would you like to see? ";
-                    size_t count = askQuestionForInt(stringStream.str());
-                    for (size_t i = 0; i < count && i < items.size(); ++i) {
-                        printMovieInfo(items[i]);
-                    }
-                }
-                else {
-                    std::cout << "\nNo movies in the database between " << startYear <<
-                              " and " << endYear << "." << std::endl;
-                }
-
-                exclusiveStartKey = result.GetResult().GetLastEvaluatedKey();
-                if (!exclusiveStartKey.empty()) {
-                    std::cout << "Not all movies were retrieved. Scanning for more."
-                              << std::endl;
-                }
-                else {
-                    std::cout << "All movies were retrieved with this scan."
-                              << std::endl;
-                }
-            }
-            else {
-                std::cout << "Failed to Scan movies: "
-                          << result.GetError().GetMessage();
-            }
-        } while (!exclusiveStartKey.empty());
-    }
-
-    // 8. Delete a movie. (DeleteItem)
-    {
-        std::stringstream stringStream;
-        stringStream << "\nWould you like to delete the movie " << title
-                     << " from the database? (y/n) ";
-        Aws::String answer = askQuestion(stringStream.str());
-        if (answer == "y") {
-            Aws::DynamoDB::Model::DeleteItemRequest request;
-            request.AddKey(YEAR_KEY, Aws::DynamoDB::Model::AttributeValue().SetN(year));
-            request.AddKey(TITLE_KEY,
-                           Aws::DynamoDB::Model::AttributeValue().SetS(title));
-            request.SetTableName(MOVIE_TABLE_NAME);
-
-            const Aws::DynamoDB::Model::DeleteItemOutcome &result = dynamoClient.DeleteItem(
-                    request);
-            if (result.IsSuccess()) {
-                std::cout << "\nRemoved \"" << title << "\" from the database."
-                          << std::endl;
-            }
-            else {
-                std::cerr << "Failed to delete the movie: "
-                          << result.GetError().GetMessage()
-                          << std::endl;
-            }
-        }
-    }
-#endif
-    askQuestion(Aws::String("End? (y/n) "));
-    // 9.Delete the table. (DeleteTable)
-    return deleteDynamoTable(MOVIE_TABLE_NAME, dynamoClient);
+     return result;
 }
 
-bool AwsDoc::DynamoDB::deleteDynamoTable(const Aws::String &tableName,
-                                         const Aws::DynamoDB::DynamoDBClient &dynamoClient) {
-    Aws::DynamoDB::Model::DeleteTableRequest request;
-    request.SetTableName(tableName);
-
-    const Aws::DynamoDB::Model::DeleteTableOutcome &result = dynamoClient.DeleteTable(
-            request);
-    if (result.IsSuccess()) {
-        std::cout << "Your Table \""
-                  << result.GetResult().GetTableDescription().GetTableName()
-                  << " was deleted!\n";
-    }
-    else {
-        std::cerr << "Failed to delete table: " << result.GetError().GetMessage();
-    }
-
-    return result.IsSuccess();
-}
-
-bool AwsDoc::DynamoDB::waitTableActive(const Aws::String &tableName,
-                                       const Aws::DynamoDB::DynamoDBClient &dynamoClient) {
-    // Repeatedly call DescribeTable until table is ACTIVE.
-    const int MAX_QUERIES = 20;
-    Aws::DynamoDB::Model::DescribeTableRequest request;
-    request.SetTableName(tableName);
-
-    int count = 0;
-    while (count < MAX_QUERIES) {
-        const Aws::DynamoDB::Model::DescribeTableOutcome &result = dynamoClient.DescribeTable(
-                request);
-        if (result.IsSuccess()) {
-            Aws::DynamoDB::Model::TableStatus status = result.GetResult().GetTable().GetTableStatus();
-
-            if (Aws::DynamoDB::Model::TableStatus::ACTIVE != status) {
-                std::this_thread::sleep_for(std::chrono::seconds(1));
-            }
-            else {
-                return true;
-            }
-        }
-        else {
-            std::cerr << "Error DynamoDB::waitTableActive "
-                      << result.GetError().GetMessage() << std::endl;
-            return false;
-         }
-        count++;
-    }
-    return false;
-}
-// snippet-end:[cpp.example_code.dynamodb.Scenario_GettingStarted]
+// snippet-end:[cpp.example_code.dynamodb.Scenario_PartiQL]
 
 
 #ifndef TESTING_BUILD
@@ -707,136 +516,21 @@ int main(int argc, char **argv) {
 
     {
         Aws::Client::ClientConfiguration clientConfig;
-        AwsDoc::DynamoDB::dynamodbGettingStartedScenario(clientConfig);
+     // snippet-start:[cpp.example_code.dynamodb.Scenario_PartiQL.main]
+     //  1. Create a table with partition: year (N) and sort: title (S). (CreateTable)
+        if (AwsDoc::DynamoDB::createDynamoDBTable(AwsDoc::DynamoDB::MOVIE_TABLE_NAME, clientConfig)) {
+
+            AwsDoc::DynamoDB::dynamodbPartiqlScenario(clientConfig);
+
+            // 9.Delete the table. (DeleteTable)
+            AwsDoc::DynamoDB::deleteDynamoTable(AwsDoc::DynamoDB::MOVIE_TABLE_NAME, clientConfig);
+        }
     }
+     // snippet-end:[cpp.example_code.dynamodb.Scenario_PartiQL.main]
 
     return 0;
 }
 
 #endif // TESTING_BUILD
 
-Aws::String AwsDoc::DynamoDB::askQuestion(const Aws::String &string,
-                                          const std::function<bool(
-                                                  Aws::String)> &test) {
-    Aws::String result;
-    do {
-        std::cout << string;
-        std::getline(std::cin, result);
-        if (result.empty()) {
-            std::cout << "Please enter some text." << std::endl;
-        }
-        if (!test(result)) {
-            continue;
-        }
-    } while (result.empty());
 
-    return result;
-}
-
-int AwsDoc::DynamoDB::askQuestionForInt(const Aws::String &string) {
-    Aws::String resultString = askQuestion(string,
-                                           [](const Aws::String &string1) -> bool {
-                                                   try {
-                                                       (void)std::stoi(string1);
-                                                       return true;
-                                                   }
-                                                   catch (const std::invalid_argument &) {
-                                                       return false;
-                                                   }
-                                           });
-
-    int result = 0;
-    try {
-        result = std::stoi(resultString);
-    }
-    catch (const std::invalid_argument &) {
-        std::cerr << "DynamoDB::askQuestionForInt string not an int "
-                  << resultString << std::endl;
-    }
-    return result;
-}
-
-float AwsDoc::DynamoDB::askQuestionForFloatRange(const Aws::String &string, float low,
-                                                 float high) {
-    Aws::String resultString = askQuestion(string, [low, high](
-            const Aws::String &string1) -> bool {
-            try {
-                float number = std::stof(string1);
-                return number >= low && number <= high;
-            }
-            catch (const std::invalid_argument &) {
-                return false;
-            }
-    });
-    float result = 0;
-    try {
-        result = std::stof(resultString);
-    }
-    catch (const std::invalid_argument &) {
-        std::cerr << "DynamoDB::askQuestionForFloatRange string not an int "
-                  << resultString << std::endl;
-    }
-
-    return result;
-}
-
-int AwsDoc::DynamoDB::askQuestionForIntRange(const Aws::String &string, int low,
-                                             int high) {
-    Aws::String resultString = askQuestion(string, [low, high](
-            const Aws::String &string1) -> bool {
-            try {
-                int number = std::stoi(string1);
-                return number >= low && number <= high;
-            }
-            catch (const std::invalid_argument &) {
-                return false;
-            }
-    });
-    int result = 0;
-    try {
-        result = std::stoi(resultString);
-    }
-    catch (const std::invalid_argument &) {
-        std::cerr << "DynamoDB::askQuestionForFloatRange string not an int "
-                  << resultString << std::endl;
-    }
-
-    return result;
-}
-
-void AwsDoc::DynamoDB::printMovieInfo(
-        const Aws::Map<Aws::String, Aws::DynamoDB::Model::AttributeValue> &movieMap) {
-    {
-        auto const &iter = movieMap.find(TITLE_KEY);
-        if (iter != movieMap.end()) {
-            std::cout << "Movie title: '" + iter->second.GetS() << "'." << std::endl;
-        }
-    }
-
-    {
-        auto const &iter = movieMap.find(YEAR_KEY);
-        if (iter != movieMap.end()) {
-            std::cout << "    Year: " + iter->second.GetN() << "." << std::endl;
-        }
-    }
-
-    {
-        auto const &iter = movieMap.find(INFO_KEY);
-        if (iter != movieMap.end()) {
-            Aws::Map<Aws::String, const std::shared_ptr<Aws::DynamoDB::Model::AttributeValue>> infoMap =
-                    iter->second.GetM();
-
-            auto const &ratingIter = infoMap.find(RATING_KEY);
-            if (ratingIter != infoMap.end()) {
-                std::cout << "    Rating: " + ratingIter->second->GetN() << "."
-                          << std::endl;
-            }
-
-            auto const &plotIter = infoMap.find(PLOT_KEY);
-            if (plotIter != infoMap.end()) {
-                std::cout << "    Synopsis: " + plotIter->second->GetS() << "."
-                          << std::endl;
-            }
-        }
-    }
-}
