+++ cpp/example_code/dynamodb/tests/dynamodb_gtests.h
@@ -31,6 +31,8 @@ namespace AwsDocTest {
 
         void AddCommandLineResponses(const std::vector<std::string>& responses);
 
+        bool createTable();
+
     private:
         static Aws::SDKOptions s_options;
 
@@ -39,6 +41,8 @@ namespace AwsDocTest {
 
         std::stringbuf  m_cinBuffer;
         std::streambuf *m_savedInBuffer = nullptr;
+
+        static bool s_TableCreated;
     };
 } // AwsDocTest
