+++ cpp/example_code/dynamodb/tests/test_main.cpp
